Fixed the mail prices [G5-18]

+++ Backend/Models/Mail/BuyedProducts.cs
@@ -10,7 +10,7 @@ namespace Backend.Models.Mail
         public string ProductName { get; set; }
         public string Description { get; set; }
         public int Amount { get; set; }
-        public int Price { get; set; }
-        public int TotalPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal TotalPrice { get; set; }
     }
 }
